model: add VideoMAE video encoders
Split TimeSformer into a separate file and PR per review.
Load the checkpoint state dict directly instead of via a lazy closure.
